Add files via upload
The project aims to create a data map of the daily food intake of an animal and analyze it's food pattern based on its intake and then compare it with the ideal intake of that animal's breed. This is done inorder to predict the reason behind the illness/abnormal behaviour of an animal based on its recent food activity. The animal taken into consideration for our project was a cow since it was one of the most prominent domestic animals in India.